testsuite: generalize ifcombine field-merge tests [PR118025]
A number of tests that check for specific ifcombine transformations
fail on AVR and PRU targets, whose type sizes and alignments aren't
conducive of the expected transformations.  Adjust the expectations.

Most execution tests should run successfully regardless of the
transformations, but a few that could conceivably fail if short and
char have the same bit width now check for that and bypass the tests
that would fail.

Conversely, one test that had such a runtime test, but that would work
regardless, no longer has that runtime test, and its types are
narrowed so that the transformations on 32-bit targets are more likely
to be the same as those that used to take place on 64-bit targets.
This latter change is somewhat obviated by a separate patch, but I've
left it in place anyway.


for  gcc/testsuite/ChangeLog

	PR testsuite/118025
	* gcc.dg/field-merge-1.c: Skip BIT_FIELD_REF counting on AVR and PRU.
	* gcc.dg/field-merge-3.c: Bypass the test if short doesn't have the
	expected size.
	* gcc.dg/field-merge-8.c: Likewise.
	* gcc.dg/field-merge-9.c: Likewise.  Skip optimization counting on
	AVR and PRU.
	* gcc.dg/field-merge-13.c: Skip optimization counting on AVR and PRU.
	* gcc.dg/field-merge-15.c: Likewise.
	* gcc.dg/field-merge-17.c: Likewise.
	* gcc.dg/field-merge-16.c: Likewise.  Drop runtime bypass.  Use
	smaller types.
	* gcc.dg/field-merge-14.c: Add comments.