ada: Compiler crash on array aggregate association iterating over function result
The compiler triggers a bug box when compiling an array aggregate with
an iterated_component_association that iterates over another array object,
failing when trying to retrieve a Choices field, which isn't an allowed
field for N_Iterated_Component_Association nodes. This occurs in procedure
Check_Function_Writable_Actuals, which wasn't accounting for the iterated
association forms.

gcc/ada/ChangeLog:

	* sem_util.adb (Check_Function_Writable_Actuals): Add handling for
	N_Iterated_Component_Association and N_Iterated_Element_Association.
	Fix a typo in an RM reference (6.4.1(20/3) => 6.4.1(6.20/3)).
	(Collect_Expression_Ids): New procedure factoring code for collecting
	identifiers from expressions of aggregate associations.
	(Handle_Association_Choices): New procedure factoring code for handling
	id collection for expressions of aggregate associations with multiple
	choices. Removed redundant test of Box_Present from original code.