final: Fix get_attr_length for asm goto [PR118411]
The problem is for inline-asm goto, the outer rtl insn type
is a jump_insn and get_attr_length does not handle ASM specially
unlike if the outer rtl insn type was just insn.

This fixes the issue by adding support for both CALL_INSN and JUMP_INSN
with asm.

OK? Bootstrapped and tested on x86_64-linux-gnu.

	PR middle-end/118411

gcc/ChangeLog:

	* final.cc (get_attr_length_1): Handle asm for CALL_INSN
	and JUMP_INSNs.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>